refactor(assistant): split browser message handler test
- move browser-originated message normalization coverage out of the main assistant handler test
- share assistant handler test setup with the Gateway provider coverage to avoid copied mock/request wiring
- reduce the Fallow handler test hotspot and duplicated-line estimate without changing runtime code
